requirements.txt
tensorflow>=2.12.0
opencv-python>=4.8.0
Pillow>=9.5.0
requests>=2.31.0
